using 'attempt' wording bc I'm not that confident

+++ mobile/android/geckoview/src/androidTest/java/org/mozilla/geckoview/test/TestCrashHandler.java
@@ -16,7 +16,12 @@ import android.os.Looper;
 import android.os.Message;
 import android.os.Messenger;
 import android.os.RemoteException;
+import java.io.ByteArrayOutputStream;
 import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
+import org.json.JSONException;
+import org.json.JSONObject;
 import org.mozilla.geckoview.GeckoRuntime;
 import org.mozilla.geckoview.test.util.UiThreadUtils;
 
@@ -102,16 +107,17 @@ public class TestCrashHandler extends Service {
      * Tests should call this to notify the crash handler that the next crash it sees is intentional
      * and that its intent should be checked for correctness.
      *
-     * @param expectFatal Whether the incoming crash is expected to be fatal or not.
+     * @param expectedProcessType The type of process the incoming crash is expected to be for.
      */
-    public void setEvalNextCrashDump(final boolean expectFatal) {
+    public void setEvalNextCrashDump(final String expectedProcessType) {
       setEvalResult(null);
       mReceiver.post(
           new Runnable() {
             @Override
             public void run() {
-              final Message msg =
-                  Message.obtain(null, MSG_EVAL_NEXT_CRASH_DUMP, expectFatal ? 1 : 0, 0);
+              final Bundle bundle = new Bundle();
+              bundle.putString(GeckoRuntime.EXTRA_CRASH_PROCESS_TYPE, expectedProcessType);
+              final Message msg = Message.obtain(null, MSG_EVAL_NEXT_CRASH_DUMP, bundle);
               msg.replyTo = mMessenger;
 
               try {
@@ -169,7 +175,7 @@ public class TestCrashHandler extends Service {
 
   private static final class MessageHandler extends Handler {
     private Messenger mReplyToMessenger;
-    private boolean mExpectFatal = false;
+    private String mExpectedProcessType;
 
     MessageHandler() {}
 
@@ -177,7 +183,8 @@ public class TestCrashHandler extends Service {
     public void handleMessage(final Message msg) {
       if (msg.what == MSG_EVAL_NEXT_CRASH_DUMP) {
         mReplyToMessenger = msg.replyTo;
-        mExpectFatal = msg.arg1 != 0;
+        Bundle bundle = (Bundle) msg.obj;
+        mExpectedProcessType = bundle.getString(GeckoRuntime.EXTRA_CRASH_PROCESS_TYPE);
         return;
       }
 
@@ -201,8 +208,8 @@ public class TestCrashHandler extends Service {
       mReplyToMessenger = null;
     }
 
-    public boolean getExpectFatal() {
-      return mExpectFatal;
+    public String getExpectedProcessType() {
+      return mExpectedProcessType;
     }
   }
 
@@ -210,6 +217,20 @@ public class TestCrashHandler extends Service {
   private MessageHandler mMsgHandler;
 
   public TestCrashHandler() {}
+
+  private static JSONObject readExtraFile(final String filePath) throws IOException, JSONException {
+    final byte[] buffer = new byte[4096];
+    final FileInputStream inputStream = new FileInputStream(filePath);
+    final ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
+    int bytesRead = 0;
+
+    while ((bytesRead = inputStream.read(buffer)) != -1) {
+      outputStream.write(buffer, 0, bytesRead);
+    }
+
+    final String contents = new String(outputStream.toByteArray(), "UTF-8");
+    return new JSONObject(contents);
+  }
 
   private EvalResult evalCrashInfo(final Intent intent) {
     if (!intent.getAction().equals(GeckoRuntime.ACTION_CRASHED)) {
@@ -222,7 +243,18 @@ public class TestCrashHandler extends Service {
 
     final File extrasFile = new File(intent.getStringExtra(GeckoRuntime.EXTRA_EXTRAS_PATH));
     final boolean extrasFileExists = extrasFile.exists();
-    extrasFile.delete();
+    try {
+      final JSONObject annotations = readExtraFile(extrasFile.getPath());
+      final String moz_crash_reason = annotations.getString("MozCrashReason");
+
+      if (!moz_crash_reason.startsWith("MOZ_CRASH(")) {
+        return new EvalResult(false, "Missing or invalid child crash annotations");
+      }
+
+      extrasFile.delete();
+    } catch (final Exception e) {
+      return new EvalResult(false, e.toString());
+    }
 
     if (!dumpFileExists) {
       return new EvalResult(false, "Dump file should exist");
@@ -232,9 +264,14 @@ public class TestCrashHandler extends Service {
       return new EvalResult(false, "Extras file should exist");
     }
 
-    final boolean expectFatal = mMsgHandler.getExpectFatal();
-    if (intent.getBooleanExtra(GeckoRuntime.EXTRA_CRASH_FATAL, !expectFatal) != expectFatal) {
-      return new EvalResult(false, "Fatality should match");
+    final String expectedProcessType = mMsgHandler.getExpectedProcessType();
+    final String processType = intent.getStringExtra(GeckoRuntime.EXTRA_CRASH_PROCESS_TYPE);
+    if (processType == null) {
+      return new EvalResult(false, "Intent missing process type");
+    }
+    if (!processType.equals(expectedProcessType)) {
+      return new EvalResult(
+          false, "Expected process type " + expectedProcessType + ", found " + processType);
     }
 
     return new EvalResult(true, "Crash Dump OK");
@@ -244,6 +281,10 @@ public class TestCrashHandler extends Service {
   public synchronized int onStartCommand(final Intent intent, final int flags, final int startId) {
     if (mMsgHandler != null) {
       mMsgHandler.reportResult(evalCrashInfo(intent));
+      // We must manually call stopSelf() here to ensure the Service gets killed once the client
+      // unbinds. If we don't, then when the next client attempts to bind for a different test,
+      // onBind() will not be called, and mMsgHandler will not get set.
+      stopSelf();
       return Service.START_NOT_STICKY;
     }
